micro-bench: factor out frequency generator helper
